+++ modules/dnn/src/layers/batch_norm_layer.cpp
@@ -13,7 +13,6 @@ Implementation of Batch Normalization layer.
 #include "layers_common.hpp"
 #include "../op_halide.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
 
 #include <opencv2/dnn/shape_utils.hpp>
