+++ modules/dnn/src/layers/fully_connected_layer.cpp
@@ -43,8 +43,7 @@
 #include "../precomp.hpp"
 #include "layers_common.hpp"
 #include "../op_halide.hpp"
-// #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
+#include "../op_inf_engine.hpp"
 
 #include <opencv2/dnn/shape_utils.hpp>
