+++ modules/dnn/src/layers/padding_layer.cpp
@@ -13,7 +13,6 @@ Implementation of padding layer, which adds paddings to input blob.
 #include "layers_common.hpp"
 #include "../op_halide.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
 
 #include <vector>
