+++ modules/dnn/src/layers/permute_layer.cpp
@@ -43,7 +43,7 @@
 #include "../precomp.hpp"
 #include "layers_common.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
+
 #include <float.h>
 #include <algorithm>
