+++ modules/dnn/src/layers/reorg_layer.cpp
@@ -51,7 +51,6 @@
 #endif
 
 #ifdef HAVE_INF_ENGINE
-#include "../ie_ngraph.hpp"
 #include <ngraph/op/experimental/layers/reorg_yolo.hpp>
 #endif
