+++ modules/dnn/src/layers/scale_layer.cpp
@@ -13,7 +13,6 @@ Implementation of Scale layer.
 #include "layers_common.hpp"
 #include "../op_halide.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
 
 #include <opencv2/dnn/shape_utils.hpp>
