+++ modules/dnn/src/layers/slice_layer.cpp
@@ -42,7 +42,6 @@
 
 #include "../precomp.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
 
 #include "layers_common.hpp"
 #include <opencv2/dnn/shape_utils.hpp>
