+++ modules/dnn/src/layers/softmax_layer.cpp
@@ -44,7 +44,6 @@
 #include "layers_common.hpp"
 #include "../op_halide.hpp"
 #include "../op_inf_engine.hpp"
-#include "../ie_ngraph.hpp"
 
 #include <algorithm>
 #include <stdlib.h>
